Reformat File Structure
