+++ gatsby-node.js
@@ -1,15 +1,15 @@
 const path = require(`path`)
-const { createFilePath } = require(`gatsby-source-filesystem`)
+// const { createFilePath } = require(`gatsby-source-filesystem`)
 
-exports.createPages = async ({ graphql, actions, reporter }) => {
-  const { createPage } = actions
-}
+// exports.createPages = async ({ graphql, actions, reporter }) => {
+//   const { createPage } = actions
+// }
 
-exports.onCreateNode = ({ node, actions, getNode }) => {
-}
+// exports.onCreateNode = ({ node, actions, getNode }) => {
+// }
 
-exports.createSchemaCustomization = ({ actions }) => {
-}
+// exports.createSchemaCustomization = ({ actions }) => {
+// }
 
 exports.onCreatePage = ({ page, actions }) => {
   const { createPage } = actions
@@ -18,5 +18,4 @@ exports.onCreatePage = ({ page, actions }) => {
 
     createPage(page)
   }
-}
-
+}
