+++ package.json
@@ -7,8 +7,10 @@
     "url": "https://github.com/openlawnz/openlawnz-website/issues"
   },
   "dependencies": {
+    "canvas": "^2.8.0",
     "gatsby": "^2.26.1",
     "gatsby-image": "^2.5.0",
+    "gatsby-plugin-alias-imports": "^1.0.5",
     "gatsby-plugin-feed": "^2.7.0",
     "gatsby-plugin-google-analytics": "^2.5.0",
     "gatsby-plugin-manifest": "^2.6.1",
