+++ src/components/hero/hero-small.jsx
