+++ src/components/layout/layout.jsx
@@ -1,10 +1,10 @@
 import React from "react"
 import PropTypes from "prop-types"
 
-import Header from "./Header"
-import Footer from "./Footer"
+import Header from "../Header/Header"
+import Footer from "../Footer/Footer"
 
-import "../css/App.css"
+import "@/css/App.css"
 
 const Layout = ({ children }) => {
   return (
